Elasticsearch: Rename public/app/plugins/datasource/elasticsearch/index_pattern.ts to follow convention
