+++ public/app/plugins/datasource/elasticsearch/IndexPattern.ts
